requirements.txt
fastapi
uvicorn
PyJWT
requests
google-genai>=1.0.0